[Refactor] Purchase 엔티티 클래스 수정
Purchase 클래스 수정

+++ src/main/java/com/nhnacademy/bookpubshop/purchase/entity/Purchase.java
@@ -41,7 +41,7 @@ public class Purchase {
     private LocalDateTime createdAt;
 
     @Column(name = "purchase_amount", nullable = false)
-    private Integer purchaseAmount;
+    private Long purchaseAmount;
 
     @ManyToOne
     @JoinColumn(name = "product_number", nullable = false)
